scoring/pic_analyze/requirements.txt
openai==1.3.0
litellm>=1.57.0
pillow==10.1.0
python-dotenv==1.0.0

